Removal of singleton directive on org.eclipse.emf.compare.
The main bundle of the framework, org.eclipse.emf.compare, declares
an extension, and hence requires
the singleton bundle directive. This makes it impossible to maintain 2
versions of the EMF Compare Framework
in the same OSGi instance.

The removed extension is a content-type contribution that is not
necessary for our applications.